% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Li6400RemarksReshuffle.R
\name{FillForward}
\alias{FillForward}
\title{Fill NA elements in a vector using the last non-NA entry without using \code{zoo:na.locf}}
\usage{
FillForward(x)
}
\arguments{
\item{x}{Vector with NA entries to be filled}
}
\value{
Returns gap filled vector
}
\description{
Fills NA elements in a vector using the last non-NA element preceeding each missing element. Follows \href{https://stat.ethz.ch/pipermail/r-help/2008-July/169195.html}{this suggestion}.
}
