% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Li6400Recalc.R
\name{Li6400Recalc}
\alias{Li6400Recalc}
\title{Recalculates Li6400 gas exchange parameters when leaf area changes using equations from Li6400 computelist}
\usage{
Li6400Recalc(data, Area = NA, inplace = FALSE)
}
\arguments{
\item{data}{data.frame containing the gas exchange data measured by a Li6400 IRGA}

\item{Area}{leaf area in [cm] to be used for the recalculation. If not provided (the default) the vector Area from data is used.}

\item{inplace}{Logical. If FALSE returns a data.frame of the recalculated gas exchange parameters only (default). If TRUE, returns the complete data.frame data with the recalculated values in place of the original values.}
}
\value{
Depending on value of argument inplace above: Data.frame with either just the re-calculated parameters or the same data.frame as data but with the recalculated values for the changed parameters.
}
\description{
Recalculates Li6400 gas exchange parameters when leaf area changes using equations from Li6400 computelist
}
