% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Li6400RemarksReshuffle.R
\name{Li6400RemarkReshuffle}
\alias{Li6400RemarkReshuffle}
\title{Re-organise remarks from a Licor 6400 IRGA measurement file previously processed via \code{\link{Li6400Import}}}
\usage{
Li6400RemarkReshuffle(remark)
}
\arguments{
\item{remark}{Data.frame with the remarks as created by the function \code{\link{Li6400Import}}}
}
\value{
Data.frame with a new vector ForwardFilledRemarks in which gaps between remarks are filled with the last, previous remark.
}
\description{
This is a rather simplistic approach to processing the between-observations remarks created by the Li6400 IRGA. Whether this approach makes sense is up to the specific use case. Assuming a block of samples (observations, Obs) is always preceeded by a remark, this approach works as the last remark is carried over to all following observations until a new remark is found. But as there can be many more remarks in file than samples and the last remark might not be the correct remark for a block of samples, the merge between the actual sample observations and remarks needs further processing afterwards. For convenience, the orginal remarks vector with gaps and the filled remarks are both provided in the output. Uses \code{\link{FillForward}}.
}
\seealso{
\code{\link{Li6400Import}}
}
