% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Li6400Import.R
\name{Li6400Import}
\alias{Li6400Import}
\title{Imports Licor 6400 photosynthesis system output files.}
\usage{
Li6400Import(file, sep = "\\t")
}
\arguments{
\item{file}{Filename of the Li6400 text file (usually .csv or .tsv)}

\item{sep}{Character string to identify columns in the Li6400 file. Default is "`\t`" for tab-separated. "," for comma-separated.}
}
\value{
Data.frame with the imported file.
}
\description{
Imports Licor 6400 photosynthesis system output files.
}
